// types.ts
export interface User {
  id: number;
  name: string;
  units: string[];
  unidade: string;
  entidade: string;
  razaoSocial: string;
  cnpj: string;
  cpf: string;
  escolaridade: string;
  genero: string;
  logradouro: string;
  dataNascimento: string;
  cargo: string;
  dataInicioVinculo: string;
  dataInicioAtividades: string;
  jornadaTrabalho: string;
  recebeAdicional: boolean;
}